refactor: more commands
