+++ packages/coln-store/Cargo.toml
@@ -31,6 +31,7 @@ rustyline = { version = "18.0.0", optional = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 shlex = "2.0.1"
+sqlparser = "0.62.0"
 thiserror = "2.0.18"
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = [
